api/requirements.txt
rdkit-pypi==2022.3.1
fastapi
uvicorn
python-multipart
